+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+ - message replies with attachments and `reply-to` functionality
+
 ### Fixes
  - Cannot use agent without the trait RemembersWhatsAppConversations